Sign shops (Very incomplete)
I can't be sure if the sign shops will work once the yml bugs are sorted
out, but the current issue I'm having is that the yml formatting does
not support custom data types, and I do not know what I need to do to
fix this. I was able to change the names system of unique signs to be a
string format, but I still can't store the data about the shop. Online
there are entries about using "Jackson" to handle yml writing, but I
don't know where the maven file for dependencies is or how to use that
in a plugin. Just like a lot of uncertainty and stuff.