Sign Shop fixes
Sign shops can no longer be destroyed by exploding tnt, creepers or
beds.
For the most part, sign shops work as they are intended to, except that
the game registers a right click with an item 2 at a time. I think the
best way to circumvent this is to recommend that the player not hold an
item while making purchases. Aside from that it works pretty much as
intended. I'm also thinking about adding a second econ wallet that can't
be used to make purchases, but can hide the amount of money you have and
serve pretty much just as a bank.