Added several untested econ commands.
BankBal: new econ wallet BANK to hide funds from the public and in a
place where the admins cannot touch them. Bankbal displays the balance
in the bank.
Deposit/Withdraw: take funds in and out of the bank and place them in
your wallet.
SetBal: Sets the balance of a player.