Steeled Resolve.
Functionally, its done except that when the server closes, players in
spectator wont be put into survival.

For the cinematics, I've got the player circling the npc twice for 20
seconds. the npc rises at the same rate the player does (2 blocks) but i
cant figure out how to get the "camera" to ignore the parts where it
clips into blocks. I don't want players to have xray vision with this.
I'm using a BlockIterator to raycast and it kind of works but it gets
really confused or something and cant figure out where to put the camera
in some places.

Really frusted rn, gonna figure it out tomorrow.