I am not good at doing this feature by feature
Added several options to wizards menu, mostly having to do with turning
off greifing from specific mobs.

Added to the different types of meat orcs can eat.

Finished magma transmutation with a small multiblock structure.

Spent a whole ass week trying to figure out what why I couldn't use
spigots loot table implementation to generate fishing loot. Turns out
the CraftLootTable file is missing a line that sets the tool for the
loot table, and for some reason fishing is the only table that needs it.
Made my own version of this class: BaugLootTable with this added line.
So much fucking nms I want to gouge out my eyeballs.

Halfway done with the last class specific skill. Actual buff is done,
and the easy half of the particle effects is done: Draws a line out of
redstone particles indication which players are recieving the buff.
Second half is going to be a rotating disk of 7 orbs that also wobbles
around the y axis. This I anticipate will be much more mathematically
intesive and take a long time. Not excited for doing it, excited for it
working.