Finished basic claiming.
Use a stick to see your available claim blocks + if the block you
clicked is available to claim.

use a shovel to claim an area.
claims can be resized.
claims cannot overlap with each other.
There is currently no function to this other than that your name is
associated with a piece of land. Similar to buying land in the
metaverse.